+++ user-service/src/main/java/com/winner/client/userservice/user/application/service/AuthCommandService.java
@@ -1,10 +1,10 @@
 package com.winner.client.userservice.user.application.service;
 
-import com.winner.client.userservice.user.application.command.LoginCommand;
-import com.winner.client.userservice.user.application.command.RefreshTokenCommand;
-import com.winner.client.userservice.user.application.command.SignupCommand;
-import com.winner.client.userservice.user.application.result.SignupResult;
-import com.winner.client.userservice.user.application.result.TokenResult;
+import com.winner.client.userservice.user.application.dto.command.LoginCommand;
+import com.winner.client.userservice.user.application.dto.command.RefreshTokenCommand;
+import com.winner.client.userservice.user.application.dto.command.SignupCommand;
+import com.winner.client.userservice.user.application.dto.result.SignupResult;
+import com.winner.client.userservice.user.application.dto.result.TokenResult;
 
 public interface AuthCommandService {
