+++ user-service/src/main/java/com/winner/client/userservice/user/application/service/UserQueryService.java
@@ -1,6 +1,6 @@
 package com.winner.client.userservice.user.application.service;
 
-import com.winner.client.userservice.user.application.result.UserDetailResult;
+import com.winner.client.userservice.user.application.dto.result.UserDetailResult;
 import java.util.UUID;
 
 public interface UserQueryService {
