+++ user-service/src/main/java/com/winner/client/userservice/user/application/service/impl/AuthCommandServiceImpl.java
@@ -4,18 +4,18 @@ import com.winner.client.global.config.jwt.JwtTokenProvider;
 import com.winner.client.global.exception.BusinessException;
 import com.winner.client.global.exception.JwtTokenErrorCode;
 import com.winner.client.userservice.common.exception.UserErrorCode;
-import com.winner.client.userservice.user.application.command.LoginCommand;
-import com.winner.client.userservice.user.application.command.RefreshTokenCommand;
-import com.winner.client.userservice.user.application.command.SignupCommand;
-import com.winner.client.userservice.user.application.result.SignupResult;
-import com.winner.client.userservice.user.application.result.TokenResult;
+import com.winner.client.userservice.user.application.dto.command.LoginCommand;
+import com.winner.client.userservice.user.application.dto.command.RefreshTokenCommand;
+import com.winner.client.userservice.user.application.dto.command.SignupCommand;
+import com.winner.client.userservice.user.application.dto.result.SignupResult;
+import com.winner.client.userservice.user.application.dto.result.TokenResult;
 import com.winner.client.userservice.user.application.service.AuthCommandService;
 import com.winner.client.userservice.user.domain.entity.User;
 import com.winner.client.userservice.user.domain.repository.UserRepository;
 import com.winner.client.userservice.user.domain.vo.Password;
 import com.winner.client.userservice.user.domain.vo.PhoneNumber;
 import com.winner.client.userservice.user.domain.vo.UserRole;
-import com.winner.client.userservice.user.infrastructure.RedisTokenRepository;
+import com.winner.client.userservice.user.infrastructure.repository.RedisTokenRepository;
 import java.util.Objects;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
