+++ user-service/src/main/java/com/winner/client/userservice/user/application/service/impl/UserQueryServiceImpl.java
@@ -2,7 +2,7 @@ package com.winner.client.userservice.user.application.service.impl;
 
 import com.winner.client.global.exception.BusinessException;
 import com.winner.client.userservice.common.exception.UserErrorCode;
-import com.winner.client.userservice.user.application.result.UserDetailResult;
+import com.winner.client.userservice.user.application.dto.result.UserDetailResult;
 import com.winner.client.userservice.user.application.service.UserQueryService;
 import com.winner.client.userservice.user.domain.entity.User;
 import com.winner.client.userservice.user.domain.repository.UserRepository;
@@ -16,6 +16,7 @@ public class UserQueryServiceImpl implements UserQueryService {
 
   private final UserRepository userRepository;
 
+  @Override
   public UserDetailResult getUserDetail(UUID userId) {
     User user = userRepository.findByIdAndDeletedAtNull(userId)
         .orElseThrow(() -> new BusinessException(UserErrorCode.USER_NOT_FOUND));
